Replace a drivers/gpu with stock code
This appears to solve the reported video recording issue.  The suspicion
is the GPU is responsible for placing timestamps, and these timestamps
are getting corrupted.  From examination I suspect LGE's code assumes the
clock on the GPU will always be on and never get out of synch.  On
startup the main chip may be very busy, so busy it is unable to handle
some frames.  From some framedrops during startup, I suspect the clock on
the GPU is kept synchronized by dropping frames, whereas
Qualcomm/LineageOS/AOSP expect the timestamps to be corrected on the main
processor or the GPU's RTC to be adjusted before recording starts.

As a result the conflict between these two strategies causes the video
corruption.  Adjusting the video stream timestamps on "corrupt" videos
should be quite possible, but this isn't my area of expertise.

I hope the change can be narrowed down further to a very small portion
(less than 50 lines), but this version isn't that big and does appear to
ameliorate the situation.